chore: Footer 분리 및 디렉토리 경로 정리
