+++ src/Components/App/App.js
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+
+function App() {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+}
+
+export default App;
